requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
fastapi==0.70.1
h11==0.12.0
idna==3.3
pycparser==2.21
pydantic==1.8.2
python-dotenv==0.19.2
requests==2.26.0
sendlk==0.0.1
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
